Fixing bug #26: added maximum allowed delta time for ephemeris (according to various GNSS constellations)

+++ src/main/java/org/gogpsproject/EphGps.java
@@ -66,21 +66,21 @@ public class EphGps implements Streamable {
 	private double iDot; /* Rate of inclination angle */
 	private double omega; /* Argument of perigee */
 	private double omega0; /*
-					 * Longitude of ascending node of orbit plane at beginning
-					 * of week
-					 */
+	 * Longitude of ascending node of orbit plane at beginning
+	 * of week
+	 */
 	private double omegaDot; /* Rate of right ascension */
 	private double M0; /* Mean anomaly at reference time */
 	private double deltaN; /* Mean motion difference from computed value */
 	private double crc, crs, cuc, cus, cic, cis; /*
-										 * Amplitude of second-order harmonic
-										 * perturbations
-										 */
-	private double fitInt; /* Fit interval */
-	
+	 * Amplitude of second-order harmonic
+	 * perturbations
+	 */
+	private long fitInt; /* Fit interval */
+
 	/* for GLONASS data */
 	private float tow;
-	
+
 	private float tauN;
 	private float gammaN;
 	private double tk;
@@ -89,7 +89,7 @@ public class EphGps implements Streamable {
 	private double Xv;
 	private double Xa;
 	private double Bn;
-	
+
 	private double Y;
 	private double Yv;
 	private double Ya;
@@ -496,139 +496,139 @@ public class EphGps implements Streamable {
 	/**
 	 * @return the fitInt
 	 */
-	public double getFitInt() {
+	public long getFitInt() {
 		return fitInt;
 	}
 	/**
 	 * @param fitInt the fitInt to set
 	 */
-	public void setFitInt(double fitInt) {
+	public void setFitInt(long fitInt) {
 		this.fitInt = fitInt;
 	}
-	
-	
+
+
 	/* for GLONASS data */
-	
+
 	public float getTauN() {
 		return tauN;
 	}
 	public void setTauN(float tauN) {
 		this.tauN = tauN;
 	}
-	
+
 	public float getGammaN() {
 		return gammaN;
 	}
 	public void setGammaN(float gammaN) {
 		this.gammaN = gammaN;
 	}
-	
+
 	public double gettk() {
 		return tk;
 	}
 	public void settk(double tk) {
 		this.tk = tk;
 	}
-	
+
 	public double getX() {
 		return X;
 	}
 	public void setX(double X) {
 		this.X = X;
 	}
-	
+
 	public double getXv() {
 		return Xv;
 	}
 	public void setXv(double Xv) {
 		this.Xv = Xv;
 	}
-	
+
 	public double getXa() {
 		return Xa;
 	}
 	public void setXa(double Xa) {
 		this.Xa = Xa;
 	}
-	
+
 	public double getBn() {
 		return Bn;
 	}
 	public void setBn(double Bn) {
 		this.Bn = Bn;
 	}
-	
+
 	public double getY() {
 		return Y;
 	}
 	public void setY(double Y) {
 		this.Y = Y;
 	}
-	
+
 	public double getYv() {
 		return Yv;
 	}
 	public void setYv(double Yv) {
 		this.Yv = Yv;
 	}
-	
+
 	public double getYa() {
 		return Ya;
 	}
 	public void setYa(double Ya) {
 		this.Ya = Ya;
 	}
-	
+
 	public int getfreq_num() {
 		return freq_num;
 	}
 	public void setfreq_num(int freq_num) {
 		this.freq_num = freq_num;
 	}
-	
+
 	public double gettb() {
 		return tb;
 	}
 	public void settb(double tb) {
 		this.tb = tb;
 	}
-	
+
 	public double getZ() {
 		return Z;
 	}
 	public void setZ(double Z) {
 		this.Z = Z;
 	}
-	
+
 	public double getZv() {
 		return Zv;
 	}
 	public void setZv(double Zv) {
 		this.Zv = Zv;
 	}
-	
+
 	public double getZa() {
 		return Za;
 	}
 	public void setZa(double Za) {
 		this.Za = Za;
 	}
-	
+
 	public double getEn() {
 		return En;
 	}
 	public void setEn(double En) {
 		this.En = En;
 	}
-	
-//	public long getEn() {
-//		return En;
-//	}
-//	public void setEn(long En) {
-//		this.En = En;
-//	}
-	
-	
+
+	//	public long getEn() {
+	//		return En;
+	//	}
+	//	public void setEn(long En) {
+	//		this.En = En;
+	//	}
+
+
 	/* (non-Javadoc)
 	 * @see org.gogpsproject.Streamable#write(java.io.DataOutputStream)
 	 */
@@ -721,7 +721,7 @@ public class EphGps implements Streamable {
 			cus = dai.readDouble();
 			cic = dai.readDouble();
 			cis = dai.readDouble();
-			fitInt = dai.readDouble();
+			fitInt = dai.readLong();
 		}else{
 			throw new IOException("Unknown format version:"+v);
 		}
